refactor: 열거형 패키지 분리
