+++ src/main/java/AttendanceApplication.java
@@ -1,0 +1,8 @@
+import controller.AttendanceController;
+
+public class AttendanceApplication {
+    public static void main(String[] args) {
+        AttendanceController attendanceController = new AttendanceController();
+        attendanceController.run();
+    }
+}
