+++ src/main/java/controller/AttendanceController.java
@@ -1,0 +1,7 @@
+package controller;
+
+public class AttendanceController {
+    public void run() {
+
+    }
+}
