+++ src/main/java/domain/attendance/Attendance.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
