+++ src/main/java/domain/attendance/AttendanceState.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.time.LocalDateTime;
