+++ src/main/java/domain/attendance/AttendanceTime.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.time.LocalDateTime;
