+++ src/main/java/domain/attendance/Holiday.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
