+++ src/main/java/util/DateTimeUtil.java
@@ -1,0 +1,9 @@
+package util;
+
+import java.time.LocalDateTime;
+
+public class DateTimeUtil {
+    public static int getDayOfWeek(LocalDateTime localDateTime) {
+        return localDateTime.getDayOfWeek().getValue();
+    }
+}
