+++ src/test/java/domain/AttendanceDateTest.java
@@ -2,6 +2,8 @@ package domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import domain.attendance.AttendanceDate;
+import domain.attendance.AttendanceState;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
